feat: add suggestions, move agent-suggestion-tab
