+++ app/components/agents/dialog-create-github-agent.tsx
@@ -15,7 +15,6 @@ import { Label } from "@/components/ui/label"
 import { createClient } from "@/lib/supabase/client"
 import { cn } from "@/lib/utils"
 import { WarningCircle, X } from "@phosphor-icons/react"
-import Link from "next/link"
 import { useRouter } from "next/navigation"
 import type React from "react"
 import { useState } from "react"
